feat: department routes,controller,model

+++ src/controllers/department_controller.ts
@@ -6,7 +6,7 @@ export const findAllDepartments =
   (DepartmentsModel: Model<Department, {}, {}, {}>) => async (req: Request, res: Response) => {
     try {
       const departments: Department[] = await DepartmentsModel.find({})
-      res.send(departments)
+      res.send({ success: true, departments })
     } catch (e) {
       res.send({ success: false, errors: e })
     }
@@ -16,7 +16,7 @@ export const createDepartment =
   (DepartmentsModel: Model<Department, {}, {}, {}>) => async (req: Request, res: Response) => {
     try {
       const newDepartment: Department = await DepartmentsModel.create({ ...req.body })
-      res.send(newDepartment)
+      res.send({ success: true, department: newDepartment })
     } catch (e) {
       res.send({ success: false, errors: e })
     }
@@ -29,7 +29,7 @@ export const updateDepartment =
         { ...req.body },
         { runValidators: true }
       )
-      res.send(updateDepartment)
+      res.send({ success: true, department: updateDepartment })
     } catch (e) {
       res.send({ success: false, errors: e })
     }
@@ -40,17 +40,20 @@ export const removeDepartment =
     const department: Department | null = await DepartmentsModel.findById(req.params.id)
     if (department) {
       const agentsFromDepartment: Agent[] = await AgentsModel.find({ department: department.name })
-      res.send(agentsFromDepartment)
+      if (agentsFromDepartment.length > 0) {
+        res
+          .status(424)
+          .send({
+            success: false,
+            errors: 'Failed to delete department. There are agents that belongs to this department',
+          })
+      }
+      if (agentsFromDepartment.length === 0) {
+        await DepartmentsModel.findByIdAndDelete(req.params.id)
+        res.send({ success: true })
+      }
     }
-    if(!department){
-      res.status(404).send('Departamento não encontrado')
+    if (!department) {
+      res.status(404).send({ success: false, errors: 'Departament not found' })
     }
-    /*try {
-      await DepartmentsModel.findByIdAndDelete(req.params.id)
-      res.send({
-        success: true,
-      })
-    } catch (e) {
-      res.send({ success: false, errors: e })
-    }*/
   }
